add motors, trapezoid profile

+++ src/main/java/frc/robot/Constants.java
@@ -31,8 +31,8 @@ public final class Constants {
     public static final double kD = 0.0;
     public static final double kI = 0.0;
     public static final double kIZone = 0.0;
-    public static final int kMaxVelocity = 65;
-    public static final int kMaxAcceleration = 200;
+    public static final double kMaxVelocity = 65.0;
+    public static final double kMaxAcceleration = 200.0;
     public static final int kMaxCurrent = 40;
